# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Thinking Spree consultant platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ai/sheet-emails:
    post:
      operationId: aiSheetEmails
      tags: [emails, ai]
      summary: Generate pre-sprint AND post-sprint email drafts from a Google Sheet link
      description: >
        Reads the linked Google Sheet (using the consultant's Sheets scope) and
        produces both a pre-sprint invitation and a post-sprint summary email.
        The consultant edits before sending — nothing auto-sends.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sheetUrl]
              properties:
                sheetUrl: { type: string, description: "Google Sheet URL or bare spreadsheet ID" }
                sprintId: { type: integer, nullable: true, description: "Optional — enriches the prompt with sprint context" }
                tone:
                  type: string
                  enum: [professional, friendly, formal, concise]
                  default: professional
      responses:
        "200":
          description: Two drafts ready for review
          content:
            application/json:
              schema:
                type: object
                properties:
                  pre:  { $ref: "#/components/schemas/EmailDraft" }
                  post: { $ref: "#/components/schemas/EmailDraft" }
                  sheetPreview: { type: string, description: "First ~600 chars of the sheet content the AI read — useful for debugging" }
  /ai/sprints/{id}/summary-update:
    post:
      operationId: aiSummaryUpdate
      tags: [emails, ai]
      summary: Propose AI-generated patches to the founder's Summary Sheet fields
      description: >
        Reads either a Google Sheet (via Sheets scope) or pasted raw notes and
        proposes a per-field patch to the founder's summary. Returns the current
        values + the proposed changes — nothing is applied automatically. The
        consultant applies via PATCH /founders/:id.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sheetUrl: { type: string, nullable: true }
                notes:    { type: string, nullable: true }
              description: "Supply at least one of sheetUrl or notes"
      responses:
        "200":
          description: Proposed updates (review before applying)
          content:
            application/json:
              schema:
                type: object
                properties:
                  founderId:    { type: integer }
                  sprintId:     { type: integer }
                  current:      { type: object, additionalProperties: true }
                  proposed:     { type: object, additionalProperties: true }
                  changedKeys:  { type: array, items: { type: string } }
...
components:
  schemas:
...
    EmailDraft:
      type: object
      properties:
        subject:
          type: string
        body:
          type: string
        to:
          type: string
        toName:
          type: string
        sprintId:
          type: integer
        emailType:
          type: string
          enum: [pre_sprint, post_sprint]
      required:
        - subject
        - body
        - to
        - toName
        - sprintId
        - emailType
...